+++ packages/pangea-sdk/pangea/asyncio/services/file_scan.py
@@ -11,9 +11,9 @@ import logging
 from typing import Dict, List, Optional, Tuple
 
 import pangea.services.file_scan as m
+from pangea import PangeaConfig
 from pangea.asyncio.request import PangeaRequestAsync
 from pangea.asyncio.services.base import ServiceBaseAsync
-from pangea.request import PangeaConfig
 from pangea.response import PangeaResponse, TransferMethod
 from pangea.utils import FileUploadParams, get_file_upload_params