+++ packages/pangea-sdk/pangea/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "6.4.0"
+__version__ = "6.5.0"
 
 from pangea.config import PangeaConfig
 from pangea.file_uploader import FileUploader