+++ packages/pangea-sdk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Made exports more explicit.
+
 ### Fixed
 
 - `PangeaResponse.http_status` being `None` when the response is not OK.